payEMIservice

+++ src/main/java/com/RWI/Nidhi/repository/LoanRepo.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface LoanRepo extends JpaRepository<Loan, Integer> {
     LoanStatus findStatusByLoanId(int loanId);
-    int findEMIByLoanId(int loanId);
-    int findFineByLoanId(int loanId);
+    double findEMIByLoanId(int loanId);
+    double findFineByLoanId(int loanId);
 }
